Made service interfaces, adjusted infos a bit

+++ gigcheck/src/main/java/priv/glycerolveinz/gigcheck/repository/gig/GigRepository.java
@@ -3,5 +3,5 @@ package priv.glycerolveinz.gigcheck.repository.gig;
 import org.springframework.data.jpa.repository.JpaRepository;
 import priv.glycerolveinz.gigcheck.model.gig.UpcomingGig;
 
-public interface GigRepository extends JpaRepository<UpcomingGig, Integer> {
+public interface GigRepository extends JpaRepository<UpcomingGig, Long> {
 }
